CourseController logic
Build two arrays "enrolledUsers" & "nonEnrolledUsers". These two arrays are filled with users who are managed by the current user who are either enrolled or nonenrolled. From here we can allow the user to make changes to these users enrollment.
If the user is an individual then they are returned as a managable user.

Currently this only supports Employers. Admins will not work.