Withdraw endpoint
Because an employer needs to be able to withdraw their employees a separate endpoint is needed to convert the form data into an enrollment class. The only other way to do this would be using JavaScript to have a dynamic action endpoint. This is neater and relys less on the clients computer working correctly.